requirements.txt
Flask
Flask-Script
Jinja2
PyYAML
